abs-ai-hub/apps/contract-reviewer/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
gradio==4.36.1
qdrant-client==1.8.2
pymupdf==1.24.2
pypdf==4.2.0
python-docx==1.1.2
ocrmypdf==16.4.0
tqdm==4.66.4
requests==2.32.3
pydantic==2.8.2
